integration_tests/requirements.txt
pytest
pymongo
pymysql
python-dotenv
pytest-cov
